+++ clang/lib/Sema/SemaStmtAsm.cpp
@@ -241,7 +241,7 @@ ExprResult Sema::ActOnGCCAsmStmtString(Expr *Expr, bool ForAsmLabel) {
 
   if (auto *SL = dyn_cast<StringLiteral>(Expr)) {
     if (!SL->isOrdinary()) {
-      Diag(SL->getBeginLoc(), diag::err_asm_string_literal_not_ordinary);
+      Diag(SL->getBeginLoc(), diag::err_asm_operand_wide_string_literal << (SL->isWide() ? 1 : 0));
       return ExprError();
     }
     if (ForAsmLabel && SL->getString().empty()) {